fix(telescope): optimize entries repository registration
Fix the storage driver registration in EntriesRepositoryFactory to properly handle multiple storage drivers and their configurations. The factory now correctly registers drivers from the telescope.storage config and instantiates the appropriate repository instance.
